Port Hamilton Nimbus to v1b1 Device/Driver/Backend architecture
Decomposes the monolithic legacy NimbusBackend into the three-layer v1b1
architecture matching the STAR port:

- Nimbus(Device): user-facing class, wires PIP capability frontend
- NimbusDriver(HamiltonTCPHandler): TCP communication, hardware discovery
- NimbusPIPBackend(PIPBackend): pipetting operations (pick_up/drop/asp/disp)
- NimbusDoor: door control subsystem (follows STARCover pattern)
- NimbusChatterboxDriver: simulation driver for testing without hardware

Also promotes the shared Hamilton TCP protocol layer from legacy to
pylabrobot/hamilton/tcp/ and converts legacy TCP submodules to thin
re-export shims (eliminating ~2,470 lines of duplication).

The legacy NimbusBackend is now a thin wrapper that delegates all
pipetting operations to NimbusPIPBackend and door operations to
NimbusDoor, maintaining the exact same API. All 157 legacy tests pass
and run through the new v1b1 code (verified via call tracing).
