+++ pylabrobot/legacy/liquid_handling/backends/hamilton/nimbus_backend_tests.py
@@ -571,8 +571,7 @@ class TestNimbusBackendUnit(unittest.IsolatedAsyncioTestCase):
       backend.set_minimum_channel_traversal_height(-10)
 
   async def test_fill_by_channels(self):
-    backend = NimbusBackend(host="192.168.1.100")
-    backend._num_channels = 8
+    backend = _setup_backend()
 
     # Test with channels 0, 2, 4
     values = [100, 200, 300]
@@ -583,8 +582,7 @@ class TestNimbusBackendUnit(unittest.IsolatedAsyncioTestCase):
     self.assertEqual(result, expected)
 
   async def test_fill_by_channels_mismatched_lengths(self):
-    backend = NimbusBackend(host="192.168.1.100")
-    backend._num_channels = 8
+    backend = _setup_backend()
 
     with self.assertRaises(ValueError):
       backend._fill_by_channels([1, 2], [0, 1, 2], default=0)
@@ -607,12 +605,24 @@ def _mock_send_command_response(command) -> Optional[dict]:
 
 def _setup_backend() -> NimbusBackend:
   """Create a NimbusBackend with pre-configured state for testing."""
+  from pylabrobot.hamilton.liquid_handlers.nimbus.door import NimbusDoor
+  from pylabrobot.hamilton.liquid_handlers.nimbus.pip_backend import NimbusPIPBackend
+
   backend = NimbusBackend(host="192.168.1.100", port=2000)
   backend._num_channels = 8
   backend._pipette_address = Address(1, 1, 257)
   backend._door_lock_address = Address(1, 1, 268)
   backend._nimbus_core_address = Address(1, 1, 48896)
   backend._is_initialized = True
+  backend._pip_backend = NimbusPIPBackend(
+    driver=backend,  # type: ignore[arg-type]
+    address=Address(1, 1, 257),
+    num_channels=8,
+  )
+  backend._door = NimbusDoor(
+    driver=backend,  # type: ignore[arg-type]
+    address=Address(1, 1, 268),
+  )
   return backend
 
 
@@ -620,6 +630,7 @@ def _setup_backend_with_deck(deck: NimbusDeck) -> NimbusBackend:
   """Create a NimbusBackend with pre-configured state and deck for testing."""
   backend = _setup_backend()
   backend._deck = deck
+  backend._pip_backend.deck = deck  # type: ignore[union-attr]
   return backend
 
 
@@ -660,6 +671,7 @@ class TestNimbusBackendCommands(unittest.IsolatedAsyncioTestCase):
 
   async def test_door_methods_without_address_raise(self):
     self.backend._door_lock_address = None
+    self.backend._door = None
 
     with self.assertRaises(RuntimeError):
       await self.backend.lock_door()
