+++ pylabrobot/legacy/liquid_handling/backends/hamilton/tcp/__init__.py
@@ -1,1 +1,31 @@
-"""Shared code for Hamilton TCP-based backends such as the Nimbus and the Prep."""
+"""Compatibility shims — canonical location is pylabrobot.hamilton.tcp."""
+
+from pylabrobot.hamilton.tcp.commands import HamiltonCommand as HamiltonCommand
+from pylabrobot.hamilton.tcp.introspection import (
+  HamiltonIntrospection as HamiltonIntrospection,
+)
+from pylabrobot.hamilton.tcp.messages import (
+  CommandMessage as CommandMessage,
+  CommandResponse as CommandResponse,
+  HoiParams as HoiParams,
+  HoiParamsParser as HoiParamsParser,
+  InitMessage as InitMessage,
+  InitResponse as InitResponse,
+  RegistrationMessage as RegistrationMessage,
+  RegistrationResponse as RegistrationResponse,
+)
+from pylabrobot.hamilton.tcp.packets import (
+  Address as Address,
+  HarpPacket as HarpPacket,
+  HoiPacket as HoiPacket,
+  IpPacket as IpPacket,
+)
+from pylabrobot.hamilton.tcp.protocol import (
+  Hoi2Action as Hoi2Action,
+  HamiltonDataType as HamiltonDataType,
+  HamiltonProtocol as HamiltonProtocol,
+  HarpTransportableProtocol as HarpTransportableProtocol,
+  HoiRequestId as HoiRequestId,
+  RegistrationActionCode as RegistrationActionCode,
+  RegistrationOptionType as RegistrationOptionType,
+)
